move TraceTest to kernel package
